Add Styling to the current HTML
I modified the h1 and h2 elements to have a color of #295f98 (a slight variation of blue), had the all the list to have the same color. I also changed the link to each recipe link to be next to the recipe's name rather than the name itself being the link where the link has the text 'View Recipe'.